# Avoid reading from Viper for path and URL funcs