--- a/hugolib/permalinks.go
+++ b/hugolib/permalinks.go
@@ -19,8 +19,6 @@ import (
 	"regexp"
 	"strconv"
 	"strings"
-
-	"github.com/spf13/hugo/helpers"
 )
 
 // pathPattern represents a string which builds up a URL from attributes
@@ -152,14 +150,14 @@ func pageToPermalinkDate(p *Page, dateField string) (string, error) {
 func pageToPermalinkTitle(p *Page, _ string) (string, error) {
 	// Page contains Node which has Title
 	// (also contains URLPath which has Slug, sometimes)
-	return helpers.URLize(p.Title), nil
+	return p.Site.pathSpec.URLize(p.Title), nil
 }
 
 // pageToPermalinkFilename returns the URL-safe form of the filename
 func pageToPermalinkFilename(p *Page, _ string) (string, error) {
 	//var extension = p.Source.Ext
 	//var name = p.Source.Path()[0 : len(p.Source.Path())-len(extension)]
-	return helpers.URLize(p.Source.TranslationBaseName()), nil
+	return p.Site.pathSpec.URLize(p.Source.TranslationBaseName()), nil
 }
 
 // if the page has a slug, return the slug, else return the title
@@ -173,7 +171,7 @@ func pageToPermalinkSlugElseTitle(p *Page, a string) (string, error) {
 		if strings.HasSuffix(p.Slug, "-") {
 			p.Slug = p.Slug[0 : len(p.Slug)-1]
 		}
-		return helpers.URLize(p.Slug), nil
+		return p.Site.pathSpec.URLize(p.Slug), nil
 	}
 	return pageToPermalinkTitle(p, a)
 }