… helpers/pathspec_test.go
… helpers/url_test.go
… hugolib/pageSort_test.go
… hugolib/page_test.go
… hugolib/pagination_test.go
… hugolib/permalinks_test.go
… tpl/template_funcs_test.go
… tpl/template_i18n_test.go