@bep 2016-09-28
The `PathSpec` receiver should be a pointer, see https://github.com/golang/go/wiki/CodeReviewComments#receiver-type (a easy-to-understand guide)