@bep 2016-10-08
I was a little chocked to see these numbers:

```
benchmark           old ns/op       new ns/op       delta
BenchmarkHugo-4     27147354215     15766184068     -41.92%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     73049380       53230076       -27.13%

benchmark           old bytes       new bytes       delta
BenchmarkHugo-4     11183517056     10158413432     -9.17%
```

But it had a good explanation ...

https://github.com/spf13/hugo/issues/2536

We had a little luck there, because I noticed @spf13 doing a Viper update right before release (got the commit right before the "slow one", which kind of defeats the purpose of vendoring ...