@bep 2016-10-08
This is the real effect comared to 0.17, which is still worth it:

```
benchmark           old ns/op       new ns/op       delta
BenchmarkHugo-4     14275127170     12655810127     -11.34%

benchmark           old allocs     new allocs     delta
BenchmarkHugo-4     55092900       48163000       -12.58%

benchmark           old bytes      new bytes      delta
BenchmarkHugo-4     9934730800     9805134368     -1.30%
```